Restructured project: moved and renamed modules, updated imports
