+++ src/test_ternary_tree.py
@@ -1,16 +1,12 @@
 import os
 import unittest
-from ternary_search_tree.ternary_search_tree import TernarySearchTree
+from ternary_search_tree import TernarySearchTree
 
 
 def load_words_from_file(relative_path: str) -> list[str]:
-    """Helper function to load words from a file relative to this test file, one per line.
-    
-    Returns:
-        list[str]: A list of non-empty stripped lines.
-    """
-    base_dir = os.path.dirname(__file__)
-    filepath = os.path.join(base_dir, "..", relative_path)
+    """Helper function to load words from a file relative to this test file."""
+    base_dir = os.path.dirname(__file__)  # Path to src/
+    filepath = os.path.join(base_dir, relative_path)  # Path to src/data/filename.txt
     with open(filepath, "r") as file:
         return [line.strip() for line in file if line.strip()]
 
@@ -19,17 +15,16 @@ class TestTernarySearchTree(unittest.TestCase):
     """
     Unit tests for the TernarySearchTree class.
     """
+
     def setUp(self):
         """
         Set up test environment with a populated TST and test word lists.
         """
         self.ternary_search_tree = TernarySearchTree()
         self.words = load_words_from_file("data/insert_words.txt")
-        self.non_existing = load_words_from_file(
-            os.path.join("data/not_insert_words.txt")
-        )
+        self.non_existing = load_words_from_file("data/not_insert_words.txt")
 
-        for word in self.words:    # Insert test words into the tree
+        for word in self.words:
             self.ternary_search_tree.insert(word)
 
     def test_insert_and_search_existing_words(self):
@@ -63,12 +58,13 @@ class TestTernarySearchTree(unittest.TestCase):
         """
         Test autocomplete results for common prefixes.
         """
-        if "banana" in self.words:
-            expected = [w for w in self.words if w.startswith("ba")]
-            self.assertCountEqual(self.ternary_search_tree.autocomplete("ba"), expected)
-        if "app" in self.words:
-            expected = [w for w in self.words if w.startswith("app")]
-            self.assertCountEqual(self.ternary_search_tree.autocomplete("app"), expected)
+        prefix = "ba"
+        expected = [w for w in self.words if w.startswith(prefix)]
+        self.assertCountEqual(self.ternary_search_tree.autocomplete(prefix), expected)
+
+        prefix = "app"
+        expected = [w for w in self.words if w.startswith(prefix)]
+        self.assertCountEqual(self.ternary_search_tree.autocomplete(prefix), expected)
 
     def test_autocomplete_no_results(self):
         """
@@ -77,5 +73,5 @@ class TestTernarySearchTree(unittest.TestCase):
         self.assertEqual(self.ternary_search_tree.autocomplete("zzz"), [])
 
 
-if __name__ == "_main_":
-   unittest.main()
+if __name__ == "__main__":
+    unittest.main()
